minor alterations so it compiles with linux

+++ NoitaChestFinderCUDA/structs/primitives.h
@@ -25,15 +25,15 @@ struct Vec2i {
 	}
 };
 
-uint64_t operator""_KB(uint64_t x)
+unsigned long long operator""_KB(unsigned long long x)
 {
 	return x * 1024;
 }
-uint64_t operator""_MB(uint64_t x)
+unsigned long long operator""_MB(unsigned long long x)
 {
 	return x * 1024 * 1024;
 }
-uint64_t operator""_GB(uint64_t x)
+unsigned long long operator""_GB(unsigned long long x)
 {
 	return x * 1024 
 		* 1024 * 1024;
